empresa/api/serializers.py
from rest_framework import serializers
from empresa.models import Empresa

class EmpresaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Empresa
        fields = '__all__'
        depth =1
