requirements.txt
django
djangorestframework
django-filter
gunicorn
django-crum==0.7.6